Move EventNewBalanceProofReceived to mediated_transfer/events.py
